requirements.txt
asgiref==3.8.1
click==8.1.7
cloudinary==1.41.0
dj-database-url==2.2.0
Django==5.0.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==22.0.0
pillow==10.4.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.5.1
svgwrite==1.4.3
Tree==0.2.4
whitenoise==6.7.0
